use long to gain more accuration

+++ src/main/java/akka/pi/WorkMessage.java
@@ -1,27 +1,27 @@
 package akka.pi;
 
 public class WorkMessage {
-    private int round;
-    private int numsElements;
+    private long round;
+    private long numsElements;
 
-    public WorkMessage(int round, int numsElements) {
+    public WorkMessage(long round, long numsElements) {
         this.round = round;
         this.numsElements = numsElements;
     }
 
-    public int getRound() {
+    public long getRound() {
         return round;
     }
 
-    public void setRound(int round) {
+    public void setRound(long round) {
         this.round = round;
     }
 
-    public int getNumsElements() {
+    public long getNumsElements() {
         return numsElements;
     }
 
-    public void setNumsElements(int numsElements) {
+    public void setNumsElements(long numsElements) {
         this.numsElements = numsElements;
     }
 }
